AVX対応のmobilityとFlipの分岐なしbit演算化

+++ src/search/search.h
@@ -30,7 +30,7 @@ void InitTree(SearchTree *tree, unsigned char midDepth, unsigned char endDepth, 
 void DeleteTree(SearchTree *tree);
 void ConfigTree(SearchTree *tree, unsigned char midDepth, unsigned char endDepth);
 void ResetTree(SearchTree *tree);
-uint64 Search(SearchTree *tree, uint64 own, uint64 opp, uint8 choiceSecond);
+uint8 Search(SearchTree *tree, uint64 own, uint64 opp, uint8 choiceSecond);
 void PVS(SearchTree *tree);
 
 #endif
